Add GitHub action to run GRChombo tests
The action has been written to build and run on the github runners which
currently are Ubuntu 18.04 machines. It is compiled with GCC v8 (v9
seemed to not work) and the OpenMPI MPI implementation. An example
Make.defs.local to build Chombo on Ubuntu 18.04 which is used by this
action has also been added.