Add parameter checking
This resolves #100 by implementing checking of parameters after they are
loaded and aborts the whole code if a check fails. Additionally, the
parameter loading and checking can be done without running the whole
simulation by passing the `-check_params` flag e.g.
`mpiexec /path/to/exec.ex /path/to/params.txt -check_params` for the
provided examples (user examples will need to be updated acoordingly).
Note that a warning function has also been added which will not abort
the code but just print a message to stderr under the provided
conditions. The checking of parameters has also been added to the
BinaryBH and KerrBH examples (not the ScalarField one has this is
changing in #140).

Note that this also fixes #136.