Speed up SurfaceExtraction (#147)
* Speed up SurfaceExtraction

Now only rank 0 queries points for interpolation and does the
integration. This significantly reduces the amount of communication the
interpolator needs to do. The only potential problem is if a user
assumes a calculated integral is available on all ranks.

* Add ability for SurfaceExtraction to broadcast integral

This is probably not very useful but given that SurfaceExtraction
computed the result on all ranks previously, this at least allows the
possibility of having the integral on all ranks without the user having
to broadcast the result themselves. This is tested in the
Spherical Extraction Test.