Add grid_buffer_size parameter (#153)
* Add additional_grid_buffer parameter

This controls the number of cells between the boundary of level l and
the boundary of level l-1. It is 'additional' because
ceil(num_ghosts/ref_ratio) is already added to it for proper nesting.

* Reparameterise additional_grid_buffer to grid_buffer_size

The new parameter is no longer added to ceil(num_ghosts/max_ref_ratio)
but is checked to ensure that it is at least this value. This should
hopefully be a little easier to understand.